clang-tidy fixes and disabling ranges for clang builds (: (#11)
clang tidy auto fixes (:

and a temporary trick for disabling ranges when building with clang or using clang tooling. these parts dont compile with clang yet anyway and I think having them disabled makes the clang tooling output nicer (:

also pointer alignment is left now :p I tried to make it that way from the beginning, but I'm not that great at using clang-format yet :p

+++ src/include/luma_av/parser.hpp
@@ -14,6 +14,7 @@ extern "C" {
 #include <luma_av/codec.hpp>
 #include <luma_av/result.hpp>
 #include <luma_av/util.hpp>
+#include <luma_av/detail/ranges_config.hpp>
 
 namespace luma_av {
 
@@ -62,7 +63,8 @@ std::pair<result<std::span<const uint8_t>>, std::span<const uint8_t>> ParseStep(
     } else if (!(data_out) or (size_out == 0)) {
         return {errc::parser_hungry_uwu, in_buff.subspan(ret)};
     } else {
-        return {std::span<uint8_t>{data_out, size_out}, in_buff.subspan(ret)};
+      return {std::span<uint8_t>{data_out, static_cast<std::size_t>(size_out)},
+              in_buff.subspan(ret)};
     }
 }
 std::pair<result<void>, std::span<const uint8_t>> ParseStep(Packet& out_pkt,
